finally fixed the issue with saving PNGs
source of the issue:
https://github.com/lastbattle/Harepacker-resurrected/commit/01873dd204aaf800e1111411ccb12786dd97e740#diff-1866cbe958d92bc2e7fa0c65aeb40b45R98

ref:
http://forum.ragezone.com/f702/release-harepacker-resurrected-1149521/index9.html#post9023856

+++ MapleLib/WzLib/WzFile.cs
@@ -362,7 +362,7 @@ namespace MapleLib.WzLib
             else
                 WzIv = WzTool.GetIvByMapleVersion(savingToPreferredWzVer); // custom selected
 
-            bool bIsWzIvSimilar = true; // checks if its saving to the same IV.
+            bool bIsWzIvSimilar = true; // check if its saving to the same IV.
             for (int i = 0; i < WzIv.Length; i++)
             {
                 if (WzIv[i] != wzDir.WzIv[i]) 
@@ -382,7 +382,7 @@ namespace MapleLib.WzLib
 
             using (WzBinaryWriter wzWriter = new WzBinaryWriter(File.Create(path), WzIv))
             {
-                wzWriter.Hash = (uint)versionHash;
+                wzWriter.Hash = versionHash;
                 Header.FSize = totalLen - Header.FStart;
                 for (int i = 0; i < 4; i++)
                 {
